requirements.txt
black==23.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
filelock==3.12.2
graphviz==0.20.1
idna==3.4
iniconfig==2.0.0
joblib==1.2.0
minimalkv==1.6.1
more-itertools==9.1.0
mypy-extensions==1.0.0
numpy==1.25.0
packaging==23.1
pandas==2.0.2
pathspec==0.11.1
platformdirs==3.8.0
pluggy==1.2.0
psycopg2-binary==2.9.6
pyarrow==12.0.1
pytest==7.4.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
requests-file==1.5.1
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
startupradar-transformers @ git+https://github.com/startupradar/transformers.git@b7844f8e283bd5a1ba42d283369bb251da3c3af0
threadpoolctl==3.1.0
tldextract==3.4.4
tqdm==4.65.0
tzdata==2023.3
uritools==4.0.1
urllib3==2.0.3
